Update package.json and fix model references in backend code

+++ unihive-backend/src/main/java/org/biog/unihivebackend/service/implementation/ClubServiceImpl.java
@@ -9,10 +9,12 @@ import org.biog.unihivebackend.model.Event;
 import org.biog.unihivebackend.model.School;
 import org.biog.unihivebackend.model.Student;
 import org.biog.unihivebackend.model.User;
+import org.biog.unihivebackend.repository.AdminRepository;
 import org.biog.unihivebackend.repository.ClubRepository;
 import org.biog.unihivebackend.repository.UserRepository;
 import org.biog.unihivebackend.service.ClubService;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -21,19 +23,48 @@ import lombok.AllArgsConstructor;
 @Service
 @AllArgsConstructor
 public class ClubServiceImpl implements ClubService {
-    
+
     private final ClubRepository clubRepository;
+    private final AdminRepository adminRepository;
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
-    
-    @PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
+
     @Override
-    public Club updateClub(UUID id, Club newclub) {
+    public Club updateClub(UUID id, Club newclub, UUID schoolId) {
         Club oldclub = clubRepository.findById(id).orElseThrow(
-            () -> new NotFoundException("Club with id " + id + " not found"));
-        User olduser = oldclub.getUser_id();
-        olduser.setEmail(newclub.getUser_id().getEmail());
-        olduser.setPassword(passwordEncoder.encode(newclub.getUser_id().getPassword()));
+                () -> new NotFoundException("Club with id " + id + " not found"));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        boolean isAdmin = authentication.getAuthorities().stream()
+                .anyMatch(role -> role.getAuthority().equals("ROLE_ADMIN"));
+        if (!isAdmin) {
+            User olduser = oldclub.getUser();
+            olduser.setEmail(newclub.getUser().getEmail());
+            olduser.setPassword(passwordEncoder.encode(newclub.getUser().getPassword()));
+            userRepository.save(olduser);
+            oldclub.setClubName(newclub.getClubName());
+            oldclub.setClubLogo(newclub.getClubLogo());
+            oldclub.setClubDescription(newclub.getClubDescription());
+            oldclub.setClubBanner(newclub.getClubBanner());
+            oldclub.setClubRating(newclub.getClubRating());
+            oldclub.setRatingCount(newclub.getRatingCount());
+            oldclub.setSchool(newclub.getSchool());
+            oldclub.setEvents(newclub.getEvents());
+            oldclub.setStudents(newclub.getStudents());
+            return clubRepository.save(oldclub);
+        }
+
+        UUID loggedInUserSchoolId = adminRepository.findByUser(((User) (authentication)
+                .getPrincipal())).orElseThrow(() -> new NotFoundException("Admin not found")).getSchool()
+                .getId();
+        if (!schoolId.equals(loggedInUserSchoolId)) {
+            throw new NotFoundException("You do not have permission to modify clubs in this school");
+        }
+
+        User olduser = oldclub.getUser();
+        olduser.setEmail(newclub.getUser().getEmail());
+        olduser.setPassword(passwordEncoder.encode(newclub.getUser().getPassword()));
         userRepository.save(olduser);
         oldclub.setClubName(newclub.getClubName());
         oldclub.setClubLogo(newclub.getClubLogo());
@@ -41,57 +72,90 @@ public class ClubServiceImpl implements ClubService {
         oldclub.setClubBanner(newclub.getClubBanner());
         oldclub.setClubRating(newclub.getClubRating());
         oldclub.setRatingCount(newclub.getRatingCount());
-        oldclub.setSchool_id(newclub.getSchool_id());
         oldclub.setEvents(newclub.getEvents());
         oldclub.setStudents(newclub.getStudents());
         return clubRepository.save(oldclub);
     }
 
-    @PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
     @Override
-    public void deleteClub(UUID id) {
+    public void deleteClub(UUID id, UUID schoolId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        UUID loggedInUserSchoolId = adminRepository.findByUser(((User) (authentication)
+                .getPrincipal())).orElseThrow(() -> new NotFoundException("Admin not found")).getSchool()
+                .getId();
+        if (!schoolId.equals(loggedInUserSchoolId)) {
+            throw new NotFoundException("You do not have permission to delete clubs in this school");
+        }
         userRepository.deleteById(clubRepository.findById(id).orElseThrow(
-            () -> new NotFoundException("Club with id " + id + " not found")
-        ).getUser_id().getId());
+                () -> new NotFoundException("Club with id " + id + " not found")).getUser().getId());
         clubRepository.deleteById(id);
     }
 
-    @PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
     @Override
-    public Club getClub(UUID id) {
+    public Club getClub(UUID id, UUID schoolId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        UUID loggedInUserSchoolId = adminRepository.findByUser(((User) (authentication)
+                .getPrincipal())).orElseThrow(() -> new NotFoundException("Admin not found")).getSchool()
+                .getId();
+        if (!schoolId.equals(loggedInUserSchoolId)) {
+            throw new NotFoundException("You do not have permission to get clubs in this school");
+        }
         return clubRepository.findById(id).orElseThrow(
-            () -> new NotFoundException("Club with id " + id + " not found")
-        );
+                () -> new NotFoundException("Club with id " + id + " not found"));
     }
 
-    @PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
     @Override
-    public List<Club> getAll() {
-        return clubRepository.findAll();
+    public List<Club> getAll(UUID schoolId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        boolean isAdmin = authentication.getAuthorities().stream()
+                .anyMatch(role -> role.getAuthority().equals("ROLE_ADMIN"));
+        UUID loggedInUserSchoolId = adminRepository.findByUser(((User) (authentication)
+                .getPrincipal())).orElseThrow(() -> new NotFoundException("Admin not found")).getSchool()
+                .getId();
+        if (!isAdmin) {
+            return clubRepository.findAll();
+        }
+        if (!schoolId.equals(loggedInUserSchoolId)) {
+            throw new NotFoundException("You do not have permission to get all clubs in this school");
+        }
+        return clubRepository.findBySchool(schoolId);
     }
 
-    @PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
     @Override
     public School getSchoolByClub(UUID id) {
         return clubRepository.findById(id).orElseThrow(
-            () -> new NotFoundException("Club with id " + id + " not found")
-        ).getSchool_id();
+                () -> new NotFoundException("Club with id " + id + " not found")).getSchool();
     }
 
-    @PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
     @Override
-    public List<Event> getEventsByClub(UUID id) {
+    public List<Event> getEventsByClub(UUID id, UUID schoolId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        UUID loggedInUserSchoolId = adminRepository.findByUser(((User) (authentication)
+                .getPrincipal())).orElseThrow(() -> new NotFoundException("Admin not found")).getSchool()
+                .getId();
+        if (!schoolId.equals(loggedInUserSchoolId)) {
+            throw new NotFoundException("You do not have permission to get events by clubs in this school");
+        }
         return clubRepository.findById(id).orElseThrow(
-            () -> new NotFoundException("Club with id " + id + " not found")
-        ).getEvents();
+                () -> new NotFoundException("Club with id " + id + " not found")).getEvents();
     }
 
-    @PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
     @Override
-    public List<Student> getFollowers(UUID id) {
+    public List<Student> getFollowers(UUID id, UUID schoolId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        UUID loggedInUserSchoolId = adminRepository.findByUser(((User) (authentication)
+                .getPrincipal())).orElseThrow(() -> new NotFoundException("Admin not found")).getSchool()
+                .getId();
+        if (!schoolId.equals(loggedInUserSchoolId)) {
+            throw new NotFoundException("You do not have permission to get followers of clubs in this school");
+        }
         return clubRepository.findById(id).orElseThrow(
-            () -> new NotFoundException("Club with id " + id + " not found")
-        ).getStudents();
+                () -> new NotFoundException("Club with id " + id + " not found")).getStudents();
     }
-    
+
 }
