+++ unihive-backend/src/main/java/org/biog/unihivebackend/service/implementation/EventServiceImpl.java
@@ -6,37 +6,68 @@ import java.util.UUID;
 import org.biog.unihivebackend.exception.NotFoundException;
 import org.biog.unihivebackend.model.Club;
 import org.biog.unihivebackend.model.Event;
+import org.biog.unihivebackend.model.User;
+import org.biog.unihivebackend.repository.AdminRepository;
 import org.biog.unihivebackend.repository.EventRepository;
 import org.biog.unihivebackend.service.EventService;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import lombok.AllArgsConstructor;
 
 @Service
 @AllArgsConstructor
-public class EventServiceImpl implements EventService{
+public class EventServiceImpl implements EventService {
 
+    private final AdminRepository adminRepository;
     private final EventRepository eventRepository;
 
-    @PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
     @Override
-    public List<Event> getAll() {
-        return eventRepository.findAll();
+    public List<Event> getAll(UUID schoolId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        boolean isAdmin = authentication.getAuthorities().stream()
+                .anyMatch(role -> role.getAuthority().equals("ROLE_ADMIN"));
+        UUID loggedInUserSchoolId = adminRepository.findByUser(((User) (authentication)
+                .getPrincipal())).orElseThrow(() -> new NotFoundException("Admin not found")).getSchool()
+                .getId();
+        if (!isAdmin) {
+            return eventRepository.findAll();
+        }
+        if (!schoolId.equals(loggedInUserSchoolId)) {
+            throw new NotFoundException("You do not have permission to get all events in this school");
+        }
+        return eventRepository.findBySchool(schoolId);
     }
 
-    @PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
     @Override
-    public Event addEvent(Event event) {
+    public Event addEvent(Event event, UUID schoolId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        UUID loggedInUserSchoolId = adminRepository.findByUser(((User) (authentication)
+                .getPrincipal())).orElseThrow(() -> new NotFoundException("Admin not found")).getSchool()
+                .getId();
+        if (!schoolId.equals(loggedInUserSchoolId)) {
+            throw new NotFoundException("You do not have permission to add events in this school");
+        }
         return eventRepository.save(event);
     }
 
-    @PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
     @Override
-    public Event updateEvent(UUID id, Event newevent) {
+    public Event updateEvent(UUID id, Event newevent, UUID schoolId) {
         Event oldevent = eventRepository.findById(id).orElseThrow(
-            () -> new NotFoundException("Event with id " + id + " not found"));
-        
+                () -> new NotFoundException("Event with id " + id + " not found"));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        UUID loggedInUserSchoolId = adminRepository.findByUser(((User) (authentication)
+                .getPrincipal())).orElseThrow(() -> new NotFoundException("Admin not found")).getSchool()
+                .getId();
+        if (!schoolId.equals(loggedInUserSchoolId)) {
+            throw new NotFoundException("You do not have permission to update events in this school");
+        }
+
         oldevent.setEventName(newevent.getEventName());
         oldevent.setEventDescription(newevent.getEventDescription());
         oldevent.setEventCategory(newevent.getEventCategory());
@@ -45,30 +76,49 @@ public class EventServiceImpl implements EventService{
         oldevent.setEventBanner(newevent.getEventBanner());
         oldevent.setEventRating(newevent.getEventRating());
         oldevent.setRatingCount(newevent.getRatingCount());
-        oldevent.setClub_id(newevent.getClub_id());
+        oldevent.setClub(newevent.getClub());
         return eventRepository.save(oldevent);
     }
 
-    @PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
     @Override
-    public void deleteEvent(UUID id) {
+    public void deleteEvent(UUID id, UUID schoolId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        UUID loggedInUserSchoolId = adminRepository.findByUser(((User) (authentication)
+                .getPrincipal())).orElseThrow(() -> new NotFoundException("Admin not found")).getSchool()
+                .getId();
+        if (!schoolId.equals(loggedInUserSchoolId)) {
+            throw new NotFoundException("You do not have permission to delete events in this school");
+        }
         eventRepository.deleteById(id);
     }
 
-    @PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
     @Override
-    public Event getEvent(UUID id) {
+    public Event getEvent(UUID id, UUID schoolId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        UUID loggedInUserSchoolId = adminRepository.findByUser(((User) (authentication)
+                .getPrincipal())).orElseThrow(() -> new NotFoundException("Admin not found")).getSchool()
+                .getId();
+        if (!schoolId.equals(loggedInUserSchoolId)) {
+            throw new NotFoundException("You do not have permission to get events in this school");
+        }
         return eventRepository.findById(id).orElseThrow(
-            () -> new NotFoundException("Event with id " + id + " not found")
-        );
+                () -> new NotFoundException("Event with id " + id + " not found"));
     }
 
-    @PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
     @Override
-    public Club getClubByEvent(UUID id) {
+    public Club getClubByEvent(UUID id, UUID schoolId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        UUID loggedInUserSchoolId = adminRepository.findByUser(((User) (authentication)
+                .getPrincipal())).orElseThrow(() -> new NotFoundException("Admin not found")).getSchool()
+                .getId();
+        if (!schoolId.equals(loggedInUserSchoolId)) {
+            throw new NotFoundException("You do not have permission to get clubs by events in this school");
+        }
         return eventRepository.findById(id).orElseThrow(
-            () -> new NotFoundException("Event with id " + id + " not found")
-        ).getClub_id();
+                () -> new NotFoundException("Event with id " + id + " not found")).getClub();
     }
-    
+
 }
